requirements.txt
alembic==1.8.1
API==0.0.7
boto3==1.24.87
Flask==2.2.2
python-decouple
Flask_JWT_Extended==4.4.4
Flask_Migrate==3.1.0
Flask_SQLAlchemy==2.5.1
python-decouple==3.6
python_bcrypt==0.3.2
SQLAlchemy==1.4.41
gunicorn
